:fire: remove non-essential parts
